Reorganizing XML classes
