+++ Xml/AdrGeneralQualifierConfig.cs
@@ -1,0 +1,34 @@
+﻿using Klyte.Addresses.ModShared;
+using System.Xml.Serialization;
+
+namespace Klyte.Addresses.Xml
+{
+    public class AdrGeneralQualifierConfig
+    {
+        private string m_namesFile;
+        private string m_qualifierFile;
+
+        [XmlAttribute("namesFile")]
+        public string NamesFile
+        {
+            get => m_namesFile; set
+            {
+                m_namesFile = value;
+                AdrShared.TriggerRoadNamingChange();
+                AdrShared.TriggerDistrictChanged();
+            }
+        }
+
+        [XmlAttribute("qualifierFile")]
+        public string QualifierFile
+        {
+            get => m_qualifierFile; set
+            {
+                m_qualifierFile = value;
+                AdrShared.TriggerRoadNamingChange();
+                AdrShared.TriggerDistrictChanged();
+            }
+        }
+    }
+}
+
